// web/src/api/generated/models/adminListChargePoints200ChargePointsItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ReloadCars API
 * API do app de pontos de recarga para veículos elétricos.
 * OpenAPI spec version: 1.0.0
 */
import type { AdminListChargePoints200ChargePointsItemSource } from './adminListChargePoints200ChargePointsItemSource'

export type AdminListChargePoints200ChargePointsItem = {
  id: string
  source: AdminListChargePoints200ChargePointsItemSource
  name: string
  /** @nullable */
  description: string | null
  /** @nullable */
  address: string | null
  latitude: number
  longitude: number
  /** @nullable */
  powerKw: number | null
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   * @nullable
   */
  pricePerKwhCents: number | null
  connectors: string[]
  /** @nullable */
  openingHours: string | null
  /** @nullable */
  attribution: string | null
  /** @nullable */
  partnerName: string | null
  /** @nullable */
  partnerId: string | null
  active: boolean
}