// web/src/api/generated/models/adminListChargePointsSource.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ReloadCars API
 * API do app de pontos de recarga para veículos elétricos.
 * OpenAPI spec version: 1.0.0
 */

export type AdminListChargePointsSource =
  (typeof AdminListChargePointsSource)[keyof typeof AdminListChargePointsSource]

export const AdminListChargePointsSource = {
  partner: 'partner',
  ocm: 'ocm',
} as const